requirements.txt
shortuuid==1.0.1
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
psycopg2==2.9.3
